- Added enemy movement

+++ core/src/de/florianbuchner/trbd/entity/component/HealthComponent.java
@@ -3,9 +3,9 @@ package de.florianbuchner.trbd.entity.component;
 import com.badlogic.ashley.core.Component;
 
 public class HealthComponent implements Component {
-    public double health;
+    public float health;
 
-    public HealthComponent(double health) {
+    public HealthComponent(float health) {
         this.health = health;
     }
 }
